Allow optional arguments
This is a breaking change that will allow optional arguments to be specified
for commands. As such, the manner in which you specify arguments needed to change
as well. Before you would specify them as part of the command name string:

    slackbot.addCommand('testing one two three', 'The test command', function() {});

now you specify them as an array:

    slackbot.addCommand(['testing', 'one', 'two', 'three'], 'The test command', function() {});

This paves the way for having commands that match multiple words instead of one.
It also allows us to specify optional arguments as:

    slackbot.addCommand(['testing', 'one', { two: 2 }], 'The test command', function() {});