Add logic to allow for configured arguments.
A configured argument is an object with one key, whose value is another
object. The nested object can specify two parameters: `default` and
`restrict`.

The `default` parameter lets users specify a default value for the
configured argument.

The `restrict` parameter lets users restrict the values that are passed
in. `restrict` can be a string, number, or array. If it is an array, the
value will be permitted if it matches any element in the array. If the
value does not match the restriction, an error will be thrown.

```javascript
slackbot.addCommand(
  'test-command',
  [{ configuredCommand: { default: 'default value', restrict: ['foo',
'bar'] } }],
  'Test command',
  function (options, callback) {
    callback(null, this.ephemeralResponse(options.args.configuredArg));
  });
```

```
/testbot test-command foo

/testbot test-command bar

/testbot test-command foobar

/testbot test-command
```